Gen-AI-With-Deep-Seek-R1-main/requirements.txt
streamlit
langchain_core
langchain_community
langchain_ollama
pdfplumber

langchain 
faiss-cpu 
python-dotenv 
docling 
langchain-docling